--- bee-honey/src/pantallaPrincipal/services/api.js
@@ -1,7 +1,0 @@
-import {create} from 'apisauce'
-
-const api = create({
-    baseURL: 'https://api.xaviaiot.com/bee_honey'
-});
-
-export default api;
