--- bee-honey/src/pantallaPrincipal/services/apiariosService.js
@@ -1,3 +1,0 @@
-import api from './api.js';
-
-export const fetchApario = () => api.get('/apiaries?beekeeper_id=xxx');
