+++ bee-honey/src/services/api.js
@@ -1,0 +1,7 @@
+// import {create} from 'apisauce';
+
+// const api = create({
+//     baseURL: 'https://api.xaviaiot.com/bee_honey'
+// });
+
+// export default api;
