+++ bee-honey/src/services/apiariosService.js
@@ -1,0 +1,3 @@
+// import api from './api.js';
+
+// export const fetchApario = () => api.get('/apiaries?beekeeper_id=xxx');
